Improvements to ProgressReporter implementation
Progress updates via Update() now just increment an atomic (vs. acquire a mutex, as the previous implementation did); this should be better for scalability.

Printing the progress bar is now done in a separate thread, which lets us continue to update the bar's elapsed time count even if the Update() hasn't been called for a while.