Align MemoryArenas to cache lines.
The SPPMIntegrator and the MLTIntegrator both have per-thread MemoryArenas; they allocate them in contiguous memory, which is a slight problem, since MemoryArenas are 80 bytes big; this leads to false sharing, which hurts performance.

Taking best of 3 runs on a 32 CPU system with the "breakfast" scene, this fix speeds up the camera and photon pass by just ~1%, but speeds up SPPM grid construction by 36% (from 1.859s to 1.367s).