Fix handling of pbrtparse header file
Previously, we had messy code in the cmake file to deal with the fact that different versions of bison use different suffixes for generated C++ header files (hh and hpp, depending). (This problem is made worse by the fact that OSX ships a 10 year old version of bison, which doesn't support the %defines directive, which would help with this.)

The solution is to rename pbrtparse.yy to pbrtparse.y, which should cause all versions of bison to emit a header named .h, which is what we want. We then rename the generated .c file to .cpp and now all platforms use the same logic for this.

Fixes issues #60 and #64.