Remove bogus ReverseMultiplyGenerator() function (and all that it implies).
To save the bit reverse after multiplying by a generator matrix, it's not
that we want to process the matrix columns in reverse order; we actually
just should reverse the bits of the uint32s for each matrix column.

Thus, removed ReverseMultiplyGenerator() and fixed all of the various
generator matrices that were mis-ordered so that they would work with that.