Add "pixelbounds" parameter to integrators (all but MLT).
Only samples within the provided bounds are rendered; the intended usage is
debugging, when a bug is happening in a known pixel or range of pixels.

(Unfortunately, the film's crop window isn't a good fit to this use case,
since the current Samplers end up generating different samples in a pixel
if the crop window is adjusted. Also, this takes integer pixel values
rather than floats in [0,1], which is more convenient for this use.)