Add rrThreshold parameter to Path and VolPath integrators.
Russian Roulette is only applied if the path contribution's luminance
is less than the threshold. This setting is useful when rendering
complex glass objects, for example, as it lets us disable Russian
roulette and keep going to the maximum path depth, which gets rid of
fine noise in very complex regions of models.