Fix rare deadlock in profiling system.
In rare circumstances, where one thread was in the middle of allocating
memory when it received the timer signal for the profiling system, and if
the thread accessed a thread-local variable for the first time in the
signal handler, it also would try to dynamically allocate memory, which
could lead to a deadlock in malloc().

Now, things are organized so that the thread-local variable, ProfilerState,
that is needed in the signal handler, is initialized in all threads before
the profiling system starts up.

Also cleaned up a bit how worker threads are launched and how they merge
their per-thread stats to the stats system.

Issue #85.