Fix obscure bug in prioritization of shape/material parameters
When creating Materials, the intent was that shape parameters override
parameters given to the material when it was originally created.
However, the way it was implemented was:

1. Shape texture parameters
2. Material texture parameters
3. Shape non-texture parameters
4. Material non-texture parameters.

Thus, given something like:

Material “foo” “texture Kd” “texmap”
Shape “bar” “rgb Kd” [ 1 2 3]

the shape override wouldn’t be used. This fixes that.